Drop unnecessary ngcp* bootoptions, change password from 'grml2011' to more generic 'grml'.
With the redesign of networking stuff and integration of eth1 we
currently do not need ngcpip{1,2}, ngcpeaddr and ngcpeiface=eth0
any longer. Instead we default to 192.168.255.25{1,2} for eth1
and use eth0 as ngcpeiface.

ngcphostname=... is not supported for pro installations since
the HA v2 setup depends on sp1 and sp2 for hostnames, therefore
do not provide the bootoption in the pro cmdlines for now.

From: Michael Prokop <mprokop@sipwise.com>