TT#15829 Ensure we run under 'set -e'
Setting options via shebang line is discouraged, as the options
don't apply if the script is executed explicitely with an
interpreter on the command line.  This is exactly what happens in
jenkins-configs' jobs/internal/grml/release.sh:

  /bin/bash -x ./build_iso.sh "private" "${GRML_ISO}"

As a result it's *not* executed under `set -e`, instead it
continuous even though there have been fatal errors already.

Change-Id: Ic9bd0afcf401e34cd536d76f4d114fbcfff0ba7c