Update patch set 5
Patch Set 5: Code-Review+1

(1 comment)

Mike, this is not final version, while something which works already.
I am going to tests and polish with Mykola, as we need to propect old releases here so new installcd will be for mr6.2+ only... 

so far your comments here are welcome!

Patch-set: 5
Label: Code-Review=+1