Update patch set 1
Patch Set 1:

I'm not sure how important this is for us, but Alex brought this up during the upgrade of Grml and whether we should expect any issues with fai-setup-storage. Since I had the basics sitting around in https://gerrit.mgm.sipwise.com/#/c/18931 already I spend a few minutes to extract the relevant bits.

I've tested the ngcp installation inside a VM (the plain Debian installation not yet), seems to work fine so far.

Please let me know if this makes sense and if it's worth it.

Patch-set: 1