Update patch set 3
Patch Set 3:

Are we comfortable with the change? Mykola, were you able to test it a bit? (FTR, I'm fine with accepting this change for master to have it available for mr6.3+, but we should be sure that we don't overlook anything that might cause us troubles once the release is out :))

PS: feel free to submit at any time yourselve

Patch-set: 3