Update patch set 3
Patch Set 3:

> IMHO submit... Mykola maybe you see some test we can add to jenkins
 > here? (to check our I/O is "expected" after installation) ?

It looks like such test is almost impossible to implement for the moment. The main problem here is that we can't separate functions' definition from the code. 'Source' or '.' executes the script so either running full script or none (aut Caesar aut nihil).

Maybe we will do it when deployment.sh is separated to smaller parts.

Patch-set: 3