Update patch set 3
Patch Set 3:

What I was asking is not a test for a code, while test for the VM state after the full installation. So we can add it into https://jenkins.mgm.sipwise.com/job/daily-build-check-infrastructure/ and run for every VM we install, we can check FS tyoe, ext4? partitions list, size, labels, ... 

While sure it is nice to have and not a super-top-priority. LGTM feel free to submit.

Patch-set: 3