TT#36454 Move HOSTNAME definition away form boot parameters loading section
Exporting HOSTNAME inside the section of loading 'boot parameters'
cause the issue if ngcphostname is not defined.
> TARGET_HOSTNAME="$(getBootParam ngcphostname)"

We have better separate loading options and using them to
have clear system behaviour, so moving the export code further.
In this case TARGET_HOSTNAME will be defined from ngcpsp1/ngcpsp2
and 'export HOSTNAME="$TARGET_HOSTNAME"' will do the job even if
ngcphostname is not defined.

We should consider to deprecate/remove 'ngcphostname' completely
if possible as currently help is a bit confusing as for me:
>  ngcphostname=... - hostname of installed system (defaults to ngcp/sp[1,2])
>                     NOTE: do NOT use when installing Pro Edition!

Change-Id: Idba31fd8286d41bd0c1aaf7ee762371e6b1d50d5