TT#35967 Check boot parameters more carefully
Function checkBootParam is fuzzy about search pattern so it
returns false positive result on similar parameters, like 'ngcpppa'
and 'ngcpppainstaller'. It leds to incorrect call of getBootParam
on nonexistent parameter so it returns 1 which causes error exit
of deployment.sh due to 'set -e'.

Add '=' symbol to the end of search pattern for parameters with argument.

Change-Id: I636a66b3e74a395de33d4bdfa9b31eb07fbe03d0