TT#37257 Use systemctl for reboot and shutdown
Use systemctl (reboot|poweroff) instead of sysrq triggers, we do not need
this workaround in a new grml.

Collect the list of running process before and after kills in the end
of deployment to catch if we do not stop them properly.

Remove outdated checks of jessie release.

Some shellcheck improvements.

Remove PROFILE variable as it does not exist/used.

Change-Id: Idda4ddf490d76e74da74d83afb919a6fa759d0e0