Update patch set 2
Patch Set 2: Code-Review+1

There's another use of this:

  2466   ps auxwww || true

Maybe it would be good to create a small function "dump_running_processes()" with the same comment that you had here, and use it in the 3 places.

But for me it's good if you want to submit as it is now.

Patch-set: 2
Reviewer: Manuel Montecelo <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+1