TT#37257 Remove kills of possible running processes
All the services which are run within chroot are stopped when they
are not needed anymore. So it makes no sense trying to kill them
by list moreover the list is outdated.

Here is the ps output before and after these kills:
https://paste.mgm.sipwise.com/show/4107/
there is no services from chroot.

Add debug output of 'systemctl status' if someone uses mountpoint
$TARGET.

Change-Id: I90f0c18157040929aaa52ec572babea1b0bf2eb1