Update patch set 1
Patch Set 1:

> > I had the impression that these were needed before ngcpcfg kicked
 > in.
 > 
 > please do not mix ngcpcfg and ngcpcfg-api, we kick the second one
 > :-)
 >
 > Since ngcpcfg care about /etc/hosts and /e/n/i on every "ngcpcfg
 > build" (which happens at the end of ngcp-installer) we do not need
 > to overwrite them later into deployment.sh
 > 
 > I am testing it as PPA now, so far installation looks good. running
 > tests now.

Hmm, no, I wasn't mixing them, it's just that some of the comments the lines that appeared in /etc/hosts and interfaces from the code just removed were not familiar (they didn't appear in the systems when I was testing other changes, as far as I can tell).

So I thought that perhaps were files written earlier during the installation process and already overwritten by ngcpcfg or ngcpcfg-api in previous versions by the time that I logged into the systems.

Anyway, as I said and gave my +1, fine if it works -- it's great to have this duplication removed, more so if it overwrites the valid files and can create problems :)

Patch-set: 1