TT#38454 Remove 'firmware-qlogic' installation case from deployment.sh
Originally the code has been added back in 2014 with commit message:
> Support blade disks we have on our internal proxmox cluster

So far at the moment the package 'firmware-qlogic' is not installed on
internal Proxmox blades (as I can see). I would consider it as
unnecessary any more and remove it here. Also it is a bit dangerous to
install some packages on internal test servers only, as production might
behave the different way then. So good to remove it here.

Change-Id: I9817ca0904eade6c51479c72f4cff428a044896f