TT#41500 Fix syncronisation of /etc/{localtime,timezone}
This stems from a bug in grml-debootstrap, https://bugs.debian.org/904607 , that
will be fixed at some point -- a fix is already committed to upstream repos by
mprokop.

Using this as a workaround:
  grml-chroot "$TARGET" dpkg-reconfigure tzdata

makes a dialog to be fired up asking for the timezone, breaking automatic
installations.

This can be possibly fixed by setting debconf selections and/or telling
dpkg-reconfigure to ask for --unseen-only, but there might be further
complications, so this is an attempt to fix it in the most straightforward way.

Change-Id: Iafd6ce123c0d3dd0f84c661d4d217c18cf075241
(cherry picked from commit 9d30670c079e338852e6c1d3b4d9af76dcee185f)