Update patch set 3
Patch Set 3:

> LGTM, unforrunately Guillem is not here, Please also fix mr6.4

 > LGTM, unforrunately Guillem is not here, Please also fix mr6.4

The reason I went with dpkg-reconfigure was because the code in the config script is doing some mapping that I didn't want to replicate. I guess if we are not going to have any of those then we could as well just use the stuff from the file as done here. The other option is to preseed debconf.

You might want to take a look at the overall logic in «dpkg-query --control-show config tzdata» BTW.

Patch-set: 3