Update patch set 3
Patch Set 3:

> > LGTM, unforrunately Guillem is not here, Please also fix mr6.4
 > 
 > The reason I went with dpkg-reconfigure was because the code in the
 > config script is doing some mapping that I didn't want to
 > replicate. I guess if we are not going to have any of those then we
 > could as well just use the stuff from the file as done here. The
 > other option is to preseed debconf.
 > 
 > You might want to take a look at the overall logic in «dpkg-query
 > --control-show config tzdata» BTW.

I looked at it a few hours ago and decided that it was complicated, so I went shopping instead.  I went shopping to mprokop's fix upstream, in particular, after your bug report :D

By the description of the issue and the fix in https://github.com/mika/grml-debootstrap/commit/a8b7e7738ac742934f213362ca3164a792e3a1ee I think that simply feeding the timezone to /etc/timezone will do, the rest of the changes in that commit are non-functional.

I suppose that the rest of the changes that dpkg-reconfigure does are not necessary in this case, because if there was no bug in grml-debootstrap, we'd be using the same code and not dpkg-reconfigure at this point in deployment anyway.

Patch-set: 3
Reviewer: Manuel Montecelo <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0