Update patch set 1
Patch Set 1:

IMHO this is two independend commits. One adds skipmediacheck + the second use drive if only one drive found. 

As for the first one LGTM without the questions... as for the second, I am not 100% sure, as there might be the second drive which was not detected properly... and currently user will abort installation as no proper drive detected in a list... after submitting, we will install on a second drive. LGTM if you really want this.

Still, please split commits :-)

Patch-set: 1