TT#28953 Ensure we have wget available by default
wget in Debian/buster:

| Package: wget
| Version: 1.19.5-1
| [...]
| Priority: standard

wget in Debian/stretch and earlier:

| Package: wget
| Version: 1.18-5+deb9u2
| [...]
| Priority: important

To ensure we still have wget available inside the
Debian system, e.g. for retrieving the ngcp-installer*.deb
file, we need to add it to the package selection.

The ngcp-nth-* packages also include wget in their dependencies,
ngcp-installer also checks and adds it, though it's too late
there for bootstrapping our systems, as we're e.g. retrieving the
installer deb + checking for PPAs using wget.

While at it move the slightly misleading comment to the
appropriate place.

Change-Id: I971b2d06528a33b1dc4c310feef83c32e10ae557