Update patch set 1
Patch Set 1:

Changes like these make us regress in the task of getting rid of ngcpcfg-api.

I was waiting in the last weeks to advance with it until the important changes of separating installation and configuration were settled, which happened in the last few days, and I don't have the full context in my head.

So it would be good to review the situation and see if we can get rid of ngcpcfg-api or not, because if we need to keep adding changes like this maybe it's not worth it.

Patch-set: 1
Label: Verified=0