Update patch set 2
Patch Set 2: Code-Review+1

> Changes like these make us regress in the task of getting rid of ngcpcfg-api...

This is true, while reality is "we cannot remove it right now".
With the current way we are installing and configuring Carrier on test Jenkins platform, there is no way to connect second+ node to trigger configuration stage.

We need a key on the server to connect it to trigger configuration which will installs keys :-(

Sure we will try to redesign soomething here, moreover we need to redesign/fix complete SSH approach we have, so far it will not happens in mr6.5. Also ngcpcfg-api will not be removed in mr6.5 as well.

I see no other way for mr6.5 right now. LGTM. ACK?

Patch-set: 2