TT#11444 Increase ROOTFS_SIZE till 10Gb
Otherwise PRO/Carrier cannot be installed because NGCP
depends on a lot of huge packages including:
 * NGCP CloudPBX firmwares ~1Gb ad for mr6.5 and growing
 * GRML squashfs ~300Mb
 * Sipwise prompts ~400Mb

At the moment of installing all the packages are downloaded
into apt cache (~2.9Gb in total as for mr6.5) + have to be
installed/unpacked (~3.2Gb in total as for mr6.5).

All-to-all I was not able to install Carrier web01a node
even with ROOTFS_SIZE=6Gb. I had to increase it till '7Gb'
to install Carrier successfully.

Lets use 10GB for root FS to have free space protection here.

The possible options were checked:
 * move apt cache into RAM, has some disadvantages, like
   we need 4GB+ RAM for cache only => hard to test on Jenkins
 * move apt cache to 'data' partition is also bad idea as
   it is a 'code' cache so have to be consistent with 'code'
   (so it is release specific and must be stored on 'root'/'fallback')

Change-Id: Ia34dcc15230e152e894e813dfbaa8688a5145de1