Update patch set 1
Patch Set 1: Code-Review+2

If the systems are going to remain installed for a long time, e.g. upgraded to future LTS versions based on Debian Buster and beyond, maybe even 10GB is not enough in the not-so-long run.

But I think that there's free space reserved to be able to increase this, and with LVM is possible to increase sizes, and 10GB should be OK for a few years, so +2 from my side.

Patch-set: 1
Reviewer: Manuel Montecelo <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2