TT#11444 Add 'nofail' mount option for '/ngcp-fallback'
NGCP software is fully functional without '/ngcp-fallback',
the boot process should NOT be aborted on the errors here.

'/ngcp-fallback' is necessary for better usability of the system
and during upgrade where we will check mount point in RW state anyway.

Also tune layout here to match with spaces count to '/ngcp-data'.

Change-Id: Ic63f87e38ee5e8582060e9ca515492274ba8c55b