TT#11444 Add RAID configuration menu
Add menu if Software RAID should be configured and which disks
should be part of it. Only 2 disks allowed.

Move all disk selection code to separate script.

Script will write variables to /tmp/disk_selection file to pass
it to main.sh. It is necessary because 'dialog' is used for
interactivity so we can not capture output of script.

Change-Id: I3dd772c9a6ac6db6809688ff38f242271d9551a8