TT#11444 Mount 'data' partition inside chroot
We perform all packages installation on 'root' partition.
Some packages, like faxserver for now, wants to create
something on 'data' partition. They do it successfully in
Debian package files '.postinst', so far the 'data'
partition have to be mounted properly otherwise the data
will gone on reboot.

Also we need to unmount it properly at the end of installation.

Change-Id: Ief90d380d71ea6e0a64eba45f403465989865952