TT#11444 Mount 'data' partition inside chroot
We perform all packages installation on 'root' partition.
Some packages, like faxserver for now, want to create
something on 'data' partition. They do it successfully in
Debian maintainer files (like postinst), so the 'data'
partition has to be mounted properly, otherwise the data
will be gone on reboot.

Also we need to unmount it properly at the end of installation.

P.S. we need to call mount under grml-chroot as chroot fails:
> (netscript.grml:1353): main():  chroot /mnt mount /ngcp-data
> mount: /ngcp-data: mount failed: Unknown error -1

Change-Id: Ief90d380d71ea6e0a64eba45f403465989865952