TT#48105 Sleep 3s to give time for the partition table to change
We avoided to do this at the time of the previous discussion when preparing
cc95e9be6219fe760c50df3078178d985155cb8f, because adding "sleep" is sometimes
prone to not be reliable and a bad idea in general.  I looked for an alternative
method and it seemed to work for a while, but now not working again.

As mentioned by Volodymyr Fedorov in the ticket, "sleep" is the most direct
solution, and the alternative is again to spend possibly a significant long time
investigating the way to get info about when the device changes (which it might
end up involving sleeps and polling the state of a device or a kernel log).

So if people are OK with it, let's do it this way now.

Change-Id: I979b9db0bec82232093c5fcc5d49291ba65901cb