Update patch set 1
Patch Set 1:

Doesn't this break every ISO that's not yet fully built by use? IMO the this check for presence of those packages is fast and doesn't cause any real overhead, while if the tools aren't there yet (e.g. because we're trying to build the deployment ISO based on an official Grml ISO for testing/debugging/...) then it will fail hard?

Patch-set: 1