Update patch set 1
Patch Set 1:

> Doesn't this break every ISO that's not yet fully built by use?

Well this code should be merge after we switch to grml-sipwise (I also support this name).

 > (e.g.
 > because we're trying to build the deployment ISO based on an
 > official Grml ISO for testing/debugging/...) then it will fail
 > hard?

Well, yes and no. I mean yes, this is fast and easy check. OTOH we are sure about the base image and it definitely contains required packages so it looks like overhead. The option of rebuilding on official grml is interesting but not sure if we need it.

Patch-set: 1
Label: Code-Review=0