Update patch set 2
Patch Set 2:

> > The option of rebuilding on official grml is interesting but not
 > sure if we need it.

It will make our life easier if we want to track down issues that are present in our internal ISO while it's known to be working in official/public Grml ISOs.
The overhead for this to just carry on the existing code is very minimal, while I'd be totally fine to remove it if we don't need in the next few months. Though we're not even pushing our own ISOs AFAICT and I'd give it some more time until we're forcing to use *only* our own self-built ISO.

 > I thought it is our current way to minimize the squashfs size

We're no longer *re*building then (but mastering/building them from scratch) :)

Patch-set: 2
Label: Code-Review=0