TT#52450 Wait until new partitions or LVM pvdevice are ready
This is the same solution for a similar problem and that it was implemented in
5ab1f5418a04fa144bec23887f2f3f0335ba092c, but extending it to other parts of the
code that did not wait-and-retry.

Apparently commands like "blockdev --flushbufs" (flush buffers of block devices)
do not solve the situation in all cases, so this is a more foolproof --if
inelegant-- solution that should not slow down the deployment more than a few
seconds at most.

Change-Id: If74e134262475ab0b100981f94fa310536f0a7ab