Update patch set 8
Patch Set 8:

> > Was just confused due to the +2 :) Alright, implemented! :)
 > 
 > I tend to +2 when even when suggesting or commenting on minor
 > things if the change is acceptable for me as it is, so if the
 > person wants the current way can go ahead without further
 > discussion/wait, and if not there will be new reviews to continue
 > discussing anyway.

(Minor because, even if the change will have a big effect, the suggestion could be disregarded and it's also a valid solution, and even if implemented it can be reverted easily if we find problems with it, etc.

But I think that it can go ahead now because for me it's better to test it for real, and revert if we find any problem, that keep wondering or discussing or adding options in many places to enable it conditionally).

Patch-set: 8