TT#56154 Add variable DEPLOYMENT_SH
This variable shows tells ngcp-installer that it is run from
deployment.sh.

In this case we need it to skip CE warning in ngcp-installer:
===================
This installation script is not intended to run in a shared system,
as it can add/delete/update existing configurations.
Please run this script only in a base install of 64 bit Debian 9 (stretch)
on a dedicated server.
Do you want to continue with the installation process? (y/N):
===================

Change-Id: I9a33ccbe07332a09a64c98a2bef844fa95498c05