Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > I think this change is bigger than we might think it is.
 > -2 to avoid submitting as-is until we're sure what we want.
 > 
 > I'd also like to see an explanation in the commit message, why you
 > think we should move from MAC address to PCI-ID? (I personally
 > don't yet understand why PCI-ID should be preferable over MAC
 > address.)

To be honest, I was mostly following suggestions and thinking that I was implementing a tried-and-true solution that we were applying by hand due to the shortcomings of the current system.

First we have the initial description and subsequent comments in https://sipwise.my.workfront.com/task/view?ID=5c76539f01052ab400611679a07e9ca7 , so it looks that the current implementation based on assigning MAC is not complete (not done in some carrier nodes, for example) and in some cases doing strange things like assigning the MAC of the underlying device to the vlan.

Then it was suggested to use the PCIID, as we/Ops are already doing in  https://wiki.sipwise.com/wiki/index.php/Ensuring_proper_VMXNET3_interface_order_on_VMWARE

In that page, the devices have different IDs, I assumed that it was "unique" as connected in the system, maybe the basis by which then we get device names like enp1s2c3.  But if there are systems in which this is not unique, sure, we cannot use it.

What we need is to get the network interfaces that the admins are using at installation time (it's not clear in my mind how they choose which is which, though), and pin them down for the future, because they will be assigned through cables/networks/switches/etc to external infrastructure, so we have to use the same network interfaces as we did initially for the lifetime of the system.

I'd say that using MAC should work fine, but not sure why Ops prefer the other method, I'll ask.

Patch-set: 2
Reviewer: Manuel Montecelo <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0