TT#61152 Add configuration Acquire::Retries=3, for apt to retry downloads
In the event of network blips, HA switchovers or failure of the servers, there
might be a problem downloading some of the hundreds or thousands of packages
needed for installation.

Set the option to retry 3 times in the case of such failures, to try to minimise
their impact.

This has to be done in several places because mmdebstrap (which uses apt) and
apt-get at invoked at different points.

Change-Id: I5acbd9895fa37452026e582c241ced945fbba3d7