Update patch set 2
Patch Set 2:

(2 comments)

> Just checking: did you test your change(s) on jenkins or so?

Probably you just wanted a YES/no, but explaining a bit in length the full problematic, because I think that you didn't follow this issue closely so maybe you have better ideas about how to address the whole issue.

I didn't test it yet because testing deployment.sh is a bit tedious, you have to copy the file somewhere and then pass it to one job installing, etc.  

If there's a problem with syntax or similar breaking existing installations we will find out soon enough (I would test it immediately), but still, maybe worth to do the tedious process to avoid this problem, so maybe I will do it.

But more importantly, because the change itself is very difficult to test.


Currently it fails only once every ~8k downloads under certain conditions (in buster, approx with glusterfs as backend, and using apt).

It is not failing with wget (at least, after attempting to download ~30k packages), so for example we don't need to do anything for "debootstrap" or other parts of our installation using wget (some of them already have --tries option or similar).  Alex and Mykola tested it with the files not being under glusterfs when serving, and also didn't get it to fail; and tried to debug it in other ways without getting a clear lead on what's the problem (AFAIK).

So, to test this script, it would have to be under these conditions (installing from Pro/Carrier second node to get the packages from approx, etc; and repeat it many times until we get to ~8k downloads, less if we are lucky, many more if we are unlucky), or to have some random failure from the server.  We could also do things like force to stop momentarily deb.sipwise.com to force a failure, but not sure if it's worth annoying our users/customers and possibly break some of our jobs running at that moment for this test.

So basically, all of these changes are designed to improve reliability.  We solved the most important problem, which was to not being able to install Carrier-trunk overnight (we couldn't for a few days, because due to the sheer amount of packages, it was always failing at the point of installing some nodes).  But this was already done by setting Acquire::Retries=3 as apt config for running system.

Changes like this one attempt to tighten even more and avoid more failures at other points like bootstrapping and installation, but it's very difficult to test them unless we force failures in the infrastructure.

Patch-set: 2