Update patch set 8
Patch Set 8: Code-Review+2

Based on conversations with Vova (he pinged me about it because it's a PITA for them to not have this working), he's been testing this for a while and it works fine for his purposes.

So I'm going to merge this now.

Patch-set: 8
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2