TT#54410 Get rid of hardcoded checksum of sipwise.gpg file
We want to get rid of old sipwise gpg file as it contains weak key. To
do it we need either update this hardcoded value (and do it every time
when key is updated) or use the same behavior as it's used in
installer.

Change-Id: Ie7d079807659c560355f2f2fbfecf4f6327f79a8
(cherry picked from commit 7bc65c5d3cd0f9f5c66500296744cc700d2d0747)