TT#54410 Get rid of hardcoded checksum of sipwise.gpg file
We want to get rid of old sipwise gpg file as it contains weak key. To
do it we need either update this hardcoded value (and do it every time
when key is updated) or use the same behavior as it's used in
installer.

Change-Id: Ie7d079807659c560355f2f2fbfecf4f6327f79a8
(cherry picked from commit 1fbbf381ced9a69daefb15cfcae80f9d368d781e)