TT#54410 Get rid of hardcoded checksum of sipwise.gpg file
We want to get rid of old sipwise gpg file as it contains weak key. To
do it we need either update this hardcoded value (and do it every time
when key is updated) or use the same behavior as it's used in
installer.

Change-Id: Id2b4055d2096372ca7b7db6d3f5a66ba644597f3
(cherry picked from commit 366f7f90cc98dab30cdac24c39cd5a6f19d10482)