TT#54410 Get rid of hardcoded checksum of sipwise.gpg file
We want to get rid of old sipwise gpg file as it contains weak key. To
do it we need either update this hardcoded value (and do it every time
when key is updated) or use the same behavior as it's used in
installer.

Change-Id: I4eba8a22cbc0b620b468fc133c07800d7f671cac
(cherry picked from commit c383a8d72dc035d7e28de42a47714162957a67d5)