TT#69519 Move functions to a single section
Previously the functions and the code were fixed so it was hard to
understand the flow of execution.
Move all functions to a single section at the beginning of the script
so it's easier to understand the code.

Fix shellcheck warnings:
SC2086: Double quote to prevent globbing and word splitting.
SC2154: efidev1 is referenced but not assigned.

Change-Id: Ie4bf28c166e4a9ff236eff807ee97adae6ecddd0