Update patch set 3
Patch Set 3: Code-Review+2

LGTM and a great idea, mixing functions in the old way was not good.

But not tested, and due to nature of bash/scripting it can be hard to exercise all code paths and detect one of the functions depends on variables not set yet.

So I'm just trusting that with tests we will eventually catch all problems craeted by this kind of refactors...

Patch-set: 3
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2