TT#105151 Run installer under "eatmydata", unless disabled by parameter
Run the installer under "eatmydata" to speed up the process.  Also add
some more information about timing.

In some VMs that we install daily ({ce,pro,carrier}-trunk.mgm) we have
the following timings:

  ce-runner, no eatmydata:
  162 seconds, 2 mins 42 secs

  ce-runner, with eatmydata:
  142 seconds, 2 mins 22 secs

  pro-runner, no eatmydata:
  246 seconds, 4 mins 06 secs

  pro-runner, with eatmydata:
  217 seconds, 3 mins 37 secs

So in these machines, for CE we save about 20 seconds, which is not much
in total but it's about 12.5% saving; and in Pro about 30 seconds (and
twice, once per machine, so about a minute in total), which is about
12.2% as well.

In Carrier, which is mostly equivalent to Pro in this respect and
typically at least 8 machines, it would mean about 4 mins in total.

When installing in hardware in previous days, maybe due to the disks
being slower, the total installation time was slightly slower:

  pro-hardware (Lenovo ThinkSystem SR250), with eatmydata:
  226 seconds, 3 mins 46 secs

Installing without eatmydata was not measured yet in hardware, but given
that the time to install is similar to the case of pro-runner, probably
the performance gain is similar too.

This looks like a relevant saving, the risk of things going wrong are
minimal, so enable it by default.

Change-Id: I8267fad08ff337c02801fb8fad0433d9b6d9f4c2
(cherry picked from commit d6b5097a8699f4ca278630a6b289285d573860ea)